Extract sub-handling and separate it from regular handling
